--- policy.go
@@ -1,6 +1,0 @@
-package main
-
-type Policy struct {
-	CarrierID    string
-	PolicyNumber string
-}
